Caliper project files added
