+++ Caliper.xcodeproj/xcuserdata/praveen.pasupuleti.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>GGKPlanDetection.xcscheme</key>
+		<key>Caliper.xcscheme</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
