+++ Caliper/Assets.xcassets/AppIcon.appiconset/Contents.json
