+++ Caliper/Assets.xcassets/Contents.json
