+++ Caliper/Assets.xcassets/img_indicator_disable.imageset/Contents.json
