+++ Caliper/Assets.xcassets/img_indicator_disable.imageset/img_indicator_disable@2x.png
