+++ Caliper/Assets.xcassets/img_indicator_disable.imageset/img_indicator_disable@3x.png
