+++ Caliper/Assets.xcassets/img_indicator_enable.imageset/Contents.json
