+++ Caliper/Assets.xcassets/img_indicator_enable.imageset/img_indicator_enable@2x.png
