+++ Caliper/Info.plist
