+++ Caliper/art.scnassets/texture.png
