--- GGKPlanDetection.xcodeproj/xcuserdata/bhanu.gollapudi.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   type = "1"
-   version = "2.0">
-</Bucket>
